Add files via upload
My objective of the analysis is to find out predictions of bike rental count on daily based on the environmental and seasonal settings, construct statistical models and then try to make prediction on rentals based on the information and models which I have. My exploration and the analysis of the data will be performed in Jupyter Notebook using Python; Tableau will be used for visualization.